+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed EllipsePanel to ChapterEllipsePanel
 - Renamed EnumerationComboBox to ChapterComboBox
 - Renamed ExtendedTreeView to ChapterTreeView
+- Renamed FormatterTextBlock to ChapterTextBlock
 ### Supported .Net Versions
 - .Net Core 3.0
 - .Net Framework 4.5